% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{LocationScaleRegressionWLS}
\alias{LocationScaleRegressionWLS}
\title{asp20plot}
\description{
The R6 class \code{asp20plot} provides parameter estimation based on
weighted/generalized least squares (WLS/GLS) and corresponding plots
}
\examples{

## ------------------------------------------------
## Method `LocationScaleRegressionWLS$new`
## ------------------------------------------------

y <- rnorm(30)
LocationScaleRegressionWLS$new(y ~ 1)

}
\section{Super class}{
\code{\link[asp20model:LocationScaleRegression]{asp20model::LocationScaleRegression}} -> \code{LocationScaleRegressionWLS}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{W}}{A numeric matrix with the weight parameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocationScaleRegressionWLS$new()}}
\item \href{#method-clone}{\code{LocationScaleRegressionWLS$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="grad_beta">}\href{../../asp20model/html/LocationScaleRegression.html#method-grad_beta}{\code{asp20model::LocationScaleRegression$grad_beta()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="grad_gamma">}\href{../../asp20model/html/LocationScaleRegression.html#method-grad_gamma}{\code{asp20model::LocationScaleRegression$grad_gamma()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="hess_beta">}\href{../../asp20model/html/LocationScaleRegression.html#method-hess_beta}{\code{asp20model::LocationScaleRegression$hess_beta()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="hess_gamma">}\href{../../asp20model/html/LocationScaleRegression.html#method-hess_gamma}{\code{asp20model::LocationScaleRegression$hess_gamma()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="loglik">}\href{../../asp20model/html/LocationScaleRegression.html#method-loglik}{\code{asp20model::LocationScaleRegression$loglik()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="resid">}\href{../../asp20model/html/LocationScaleRegression.html#method-resid}{\code{asp20model::LocationScaleRegression$resid()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionWLS$new(
  location,
  scale = ~1,
  data = environment(location),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{location}}{A two-sided formula with the response variable on the
LHS and the predictor for the location (i.e. the mean)
on the RHS.}

\item{\code{scale}}{A one-sided formula with the predictor for the scale
(i.e. the standard deviation) on the RHS.}

\item{\code{data}}{A data frame (or list or environment) in which to evaluate
the \code{location} and \code{scale} formulas.}

\item{\code{...}}{Passed on to \code{\link[stats:model.matrix]{stats::model.matrix()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \code{LocationScaleRegressionWLS} object.
}

\subsection{Returns}{
A \code{LocationScaleRegressionWLS} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- rnorm(30)
LocationScaleRegressionWLS$new(y ~ 1)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionWLS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
