% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{plot.LocationScaleRegressionWLS}
\alias{plot.LocationScaleRegressionWLS}
\title{Plot function for the \code{LocationScaleRegressionWLS} model class}
\usage{
\method{plot}{LocationScaleRegressionWLS}(model)
}
\arguments{
\item{model}{A \code{\link{LocationScaleRegressionWLS}} object.}

\item{parameters}{Plotting parameters can be set here.}

\item{summary}{Prints a summary statistics of the results.}
}
\value{
Returns ... plots and prints summary statistics if \code{summary=T}
}
\description{
This function provides plots for checking the results of a fitted
LocationScaleRegressionWLS object.
}
\examples{
y <- rnorm(30)
model <- LocationScaleRegressionWLS$new(y ~ 1)
newton_raphson_WLS(model)
plot(model)

}
